Improve error detection, update README, update copyright years
The fetch() method now returns CURLE_GOT_NOTHING when no search result is found for the given search term.
Updated the content in README.md to reflect the changes in source code and also added some comments in the code examples
Also added 2022 to the copyright notices in the source code and LICENSE.txt. Happy New Year!