Replace Rule of 0 compliance with Rule of 3
Friendship ended with Rule of 0, now Rule of 3 is my best friend
Jokes aside though, the Rule of 0 plan didn't quite work out. The unique_ptr for the curl pointer marks the copy constructor and copy assignment operator as deleted, which I don't want. Fixed it by returning to monke (raw pointer for curl) and following the Rule of 3
Bonus: improved code comments